master_auth/requirements.txt
django
psycopg2
cryptography
djangorestframework
django-cqrs
djangorestframework-simplejwt
pyjwt[crypto]
